0.4.0.0
Added a TruthTable data type to NodeGate, which represents the gate as map from input vectors to output states. Not having to convert larges table to NodeExpr format means not having to check large NodeExpr against every possible input when comparing to a logical expression from the dmms, or when generating tt files. This speeds up parsing and tt generation by a factor of 30 for large dmms files.

Signed-off-by: Peter Regan <peter.regan@gmail.com>