0.5.3.2
Bugfix - when dynmod writes a Booleannet file, it does it correctly as long as the logic expression is specified in the dmms file. If it’s autogenerated from the table, the expression written out is incorrect; it contains no “not” operators at all. Also writes gates that started out as only TruthTables more compactly.

Also, remove tt files each time dynmod -t is call, to prevent old gates from remaining.

Signed-off-by: Peter Regan <pregan@Dynamical-Modularity-Laptop.local>