0.5.3.4
Adjusted the dmms comparison function so that if, for example, there are no differences between two nodes, the name of that node is not listed. Generally eliminated text that confirmed sameness. Also, if a node name is shared bwtween two layers, but the inputs to it are different, I display the two gates for comparison’s sake, as long as they were originally logical expressions, as logical expressions generated from truth tables are useless messess.

Signed-off-by: Peter Regan <peter.regan@gmail.com>